@wip
got a troublesome test working, so this is a checkpoint. Here's the
problem: Liquid 4.0 introduced some changes, in particular the removal
of context from the Filesystem classes. I _think_ we can get the same
capability out of panoramic, so it may be a thing that only affects our
tests. We'll have to test carefully.

Next, the same undocumented behavior around partial inclusion with a passed
array popped up again in the specs. I submitted a PR but it doesn't look
like it's going anywhere.