fixup
update ResolverSystem to work with Liquid 4

Since Liquid 4 doesn't pass in a context to the file system any more, we
build our own context and pass it in. The important thing we need to
know is what controller the call is coming from, since that affects our
lookup path for templates.

I'm not happy that we have to add a parameter to the regular FileSystem
signature, but I don't see a way around it.