Fix failing specs on TemplateHandler
I'm not sure exactly when, but we have accumulated some failing specs
that need to be corrected before we merge any new features.
* the Drop class was unable to proceed if a Drop had no
attributes (which is the case for the base Drop class under test)
* The {% for %} tag exhibited the undocumented behavior when passed an
array, which is fixed by referring to my own fork of liquid.
* Liquid 3.0.0 does not have whitespace control, so the spec is modified
to account for expected whitespace.
* I added a Drop class to the dummy app, since it did not seem to behave
properly without explicitly making the Widget#name attribute visible.